+++ src/main/java/com/telepathicgrunt/the_bumblezone/fluids/HoneyFluid.java
@@ -210,7 +210,12 @@ public abstract class HoneyFluid extends ForgeFlowingFluid {
         int currentFluidLevel = blockState.getBlock() instanceof HoneyFluidBlock ? blockState.getValue(HoneyFluidBlock.LEVEL) : 0;
         int highestNeighboringFluidLevel = currentFluidLevel;
         int neighboringFluidSource = 0;
-        boolean hadAboveFluid = blockState.getBlock() instanceof HoneyFluidBlock ? blockState.getValue(ABOVE_FLUID) : false;
+        boolean hasAboveFluid = blockState.getBlock() instanceof HoneyFluidBlock ? blockState.getValue(ABOVE_FLUID) : false;
+
+        BlockPos aboveBlockPos = blockPos.above();
+        BlockState aboveBlockState = worldReader.getBlockState(aboveBlockPos);
+        BlockState belowBlockState = worldReader.getBlockState(blockPos.below());
+        boolean canPassThroughBelow = ((FlowingFluidAccessor)this).thebumblezone_callCanPassThroughWall(Direction.DOWN, worldReader, blockPos, blockState, blockPos.below(), belowBlockState);
 
         for(Direction direction : Direction.Plane.HORIZONTAL) {
             BlockPos sideBlockPos = blockPos.relative(direction);
@@ -222,21 +227,19 @@ public abstract class HoneyFluid extends ForgeFlowingFluid {
                 }
 
                 highestNeighboringFluidLevel = Math.max(highestNeighboringFluidLevel, sideFluidState.getAmount());
-                if(sideFluidState.getType() instanceof HoneyFluid) {
+                if(sideFluidState.is(BzFluidTags.BZ_HONEY_FLUID) && !(canPassThroughBelow && !sideFluidState.isSource() && sideBlockState.getValue(HoneyFluidBlock.FALLING) && aboveBlockState.getFluidState().is(BzFluidTags.BZ_HONEY_FLUID))) {
                     lowestNeighboringFluidLevel = Math.min(lowestNeighboringFluidLevel, sideFluidState.isSource() ? 0 : sideFluidState.getValue(BOTTOM_LEVEL));
                 }
             }
         }
 
-        BlockPos aboveBlockPos = blockPos.above();
-        BlockState aboveBlockState = worldReader.getBlockState(aboveBlockPos);
         FluidState aboveFluidState = aboveBlockState.getFluidState();
         boolean aboveFluidIsThisFluid = !aboveFluidState.isEmpty() && aboveFluidState.getType().isSame(this);
         int newBottomFluidLevel = Math.max(lowestNeighboringFluidLevel - 1, 0);
         boolean isFalling = true;
         int newFluidLevel = 8;
         int dropOffValue = this.getDropOff(worldReader);
-        if(hadAboveFluid && !aboveFluidIsThisFluid) {
+        if(hasAboveFluid && !aboveFluidIsThisFluid) {
             dropOffValue = 0;
         }
 
@@ -246,8 +249,7 @@ public abstract class HoneyFluid extends ForgeFlowingFluid {
             }
         }
         else {
-            BlockState belowBlockState = worldReader.getBlockState(blockPos.below());
-            isFalling = aboveFluidState.isEmpty() && neighboringFluidSource == 0 && highestNeighboringFluidLevel <= currentFluidLevel && ((FlowingFluidAccessor)this).thebumblezone_callCanPassThroughWall(Direction.DOWN, worldReader, blockPos, blockState, blockPos.below(), belowBlockState);
+            isFalling = aboveFluidState.isEmpty() && neighboringFluidSource == 0 && highestNeighboringFluidLevel <= currentFluidLevel && canPassThroughBelow;
             newFluidLevel = highestNeighboringFluidLevel - dropOffValue;
         }
 
@@ -323,15 +325,6 @@ public abstract class HoneyFluid extends ForgeFlowingFluid {
         }
 
         return totalHeight / (float)checkedSides;
-    }
-
-    // Used in FluidRendererMixin
-    public static double setBottomFluidHeight(double bottomY, BlockPos blockPos, FluidState fluidState) {
-        if(fluidState.is(BzFluidTags.BZ_HONEY_FLUID)) {
-            double blockY = (blockPos.getY() & 15);
-            return blockY + (fluidState.isSource() ? 0f : fluidState.getValue(HoneyFluidBlock.BOTTOM_LEVEL) / 8f);
-        }
-        return bottomY;
     }
 
     public static boolean shouldNotCullSide(BlockGetter world, BlockPos blockPos, Direction direction, FluidState currentFluidState) {